Make /task/public and /course/common/public ro-accessible in student
From the security point of view both 'public' folders are accessible to
any student anyways (in fact, they can be accessed even by unregistered
anonymous users).

Mounting them as read-only ensures that the student cannot modify them.

The reason for this changeset is to allow to put in public/ things like
scripts or datasets needed during the execution of the tests (in the
student container) while making them accessible to the students
(people) so they can download and use them for testing in their own
local environments.