Augment inginious conf from env vars
INGINIOUS_ALLOW_REGISTRATION controls if the registration of new users
via the native INGInious webapp is allowed or not. In testing it is
handy to set this to '1' to add as many users as you want.

In production it should be '0' so only Github accounts can be
registered.

In any case the INGINIOUS_GITHUB_AUTH_CLIENT_ID_AND_SECRET defines the
client id and client secret for the Github OAuth app that will validate
the users.